requirements.txt
pypdf
pymupdf
beautifulsoup4
pandas
streamlit
nltk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymysql
plotly
sqlalchemy
pymongo
pyarrow
spacy
boto3
presidio-analyzer
azure-storage-blob
google-cloud-storage
pytesseract
Pillow
fastavro
gliner
slack_sdk
atlassian-python-api